+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - environment variables now correctly take precedence over config file values
 
+
 ## [1.6.0] - 2026-01-05
 
 ### Added