+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.8.0] - 2026-01-22
+
 ### Fixed
 
 - add TTY checks before interactive prompts to prevent hanging in non-interactive mode
@@ -329,7 +331,8 @@
 - adds a -t, --title flag to the `issue pr` command, allowing you to provide a PR title that is different than linear's issue title
 - allows linear issue identifiers to be passed in as arguments to the issue commands as an alternative to parsing the branch name, e.g. `linear issue show ABC-123`
 
-[Unreleased]: https://github.com/schpet/linear-cli/compare/v1.7.0...HEAD
+[Unreleased]: https://github.com/schpet/linear-cli/compare/v1.8.0...HEAD
+[1.8.0]: https://github.com/schpet/linear-cli/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/schpet/linear-cli/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/schpet/linear-cli/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/schpet/linear-cli/compare/v1.4.0...v1.5.0